Remove sanitizeUrl function
The sanitation that `sanitizeUrl` offers is minimal. It returns
`about:blank` for invalid URLs. `sanitizeUrl` is used only once.

Internally, `sanitizeUrl` uses `new URL(...)` to decide if a URL is
valid. It is very unlikely that a URL like
`/${this.projectPath}/-/error_tracking/${errorId}.json` won't be valid.

If the sanitation would have been proven more useful it would have been
adopted broadly.