Rename isSafeURL for isValidURL
`isSafeURL` doesn't check if a URL is safe (harmful or malicious). It
checks if the URL is of the correct type, has the right format and
protocols. Internally, the correct URL format is validated using the
`URL` constructor, which raises an `Invalid URL` TypeError. Hence the
name `isValidURL` seems to be more adequate.