Optimize variables on pipeline creation for yaml and node
On pipeline creation, a stub build is created for every job candidate
to run rules against it. This is an expensive operation for many
reasons. More details are at
https://gitlab.com/gitlab-org/gitlab/-/issues/350057.

One of the downsides of using a stub build is that variable calculation
becomes also expensive because it tries to reach sub-relations,
for example "metadata" in the scope of this change.

Eventually, our goal is to not use a stub build at all, but I'd like to
gradually lose the connection between the AR object and variables. And,
this is the first step; we are generating yaml_variables and
node_variables from the hash instead of the AR object.

More information about the performance optimization:
https://gitlab.com/groups/gitlab-org/-/epics/14348#note_1982576223

This change is behind the feature flag
`ci_variables_optimization_for_yaml_and_node`.