Work around deadlocks in CI in admin_settings_spec.rb
It appears that on occasion Rails can get into a deadlock when
attempting to clear the query cache from both the main and CI database
(https://github.com/rails/rails/issues/45994). This was happening in
`ee/spec/features/admin/admin_settings_spec.rb` because a project was
created while the browser sent Ajax requests to the API, causing an
update to the users activity while the project attempted to create a
CI project mirror.

To work around the issue, fix the test to create the objects before
loading the page.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/471622