Parse the package.json file entirely for NPM uploads
When processing an NPM package publication, we parse the package.json
file.
Previously, this file was parsed using an event based parser. This led
to sub fields being read for top level fields.
The event based parser is replaced by the standard json parsing. We
already limit the file size that we allow for package.json, so it is
safe to parse it entirely. This way, we get an accurate json parsing.

Changelog: fixed