Merge Request auto close issues now checked on merge
Before we would not create merge_requests_closing_issues
in the DB if the referenced issues's project had
autoclose_referenced_issues disabled. Now we create the records in
the DB even if it is disabled, but we check that the setting is
enabled on the project before closgin each issue when the MR is merged.

Changelog: changed