Allow code review bot to publish review on private projects
When we execute duo code review on a private project, we call
`DraftNotes::PublishService` and it checks if user can `create_note`.
We only want to allow that if the executing user is allowed to
create note in a private project.

To do that, we check the permission of the executing user instead.
If it's allowed to do so, we allow the publishing of review. If
not, we don't publish the review.