Fix pipeline subscription error flash message display
If a pipeline subscription failed to create, the errors
returned can be an array of strings. The flash alert
doesn't seem to render the strings in
the array, so the flash alert contains blank text.

This change joins the array of strings to a single string.
This approach is used elsewhere in the codebase.

Changelog: fixed
Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/470442
EE: true