Remove obsolete sidebar_assignees component
And related code. Currently we have _two_ components (at least) for
rendering an Assignee Widget in an issuable sidebar.

The more modern one `sidebar_assignees_widget.vue` is rendered if:
```
isInIssuePage() || isInDesignPage() || isInMRPage()
```

Which is all the pages with Issuables that have Assignees. Incidents
seem to be "Issue" pages as well. Therefore we can remove the "else"
code path and all the related code.