Migrate os sbom occurrences to correct components
This migration corrects the components, and their occurrences that were
left over from the '20240425205205' migration. It does two things:

1. It updates the "sbom_occurrences".component_name so that it no longer
   uses the os prefix e.g. 'alpine/curl' becomes 'curl'.
2. It updates the "sbom_occurrences".component_id so that it points to
   the correct parent component.

This ensures that we can navigate to the correct parent component using
active record calls in the future, and have the correct name in the
denormalized column. This intentionally does **not** migrate any
vulnerability data! The reason for this is three fold:

1. It reduces the complexity of the migration which means the migration
   is less of a risk.
2. We put less burden on the database by writing to fewer tables.
3. The findings will be corrected in future scans, and the previous
   findings that don't match up will be marked as no longer found. These
   then can be remediated in mass, and is much simpler to do versus
   writing a batched background migration.

Changelog: changed