Bulk insert CVS vulnerability scanners
This improves the efficiency of scans by removing an N+1 query issue
that was adopted in the transition to global type scans (scans that
cover many projects at once). Previously, we tried to upsert scanners N
times (N being the number of projects in a scan), and as a result put a
lot of unnecessary pressure on the database. This should reduce the WAL
in theory, and overall improve the performance of the scanner.

Changelog: performance