Make cursor tracker redis key include filename
To support running multiple rake tasks in parallel based on separate
list files, we want to make the cursor tracker redis key include the
filename of the list it is processing. This way we will have a way to
resume correctly if ever one of the rake tasks gets interrupted.