Add an API for querying if a user can edit biling
We need to be able to determine the permission level of the user on
CustomersDot, to verify if they are allowed to make a purchase for a
given namespace.

Introduce a new API that will allow us to make this query based on the
existing group policy, rather than duplicating the logic in
CustomersDot.

Issue: https://gitlab.com/gitlab-org/customers-gitlab-com/-/issues/9773
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/159176