Remove pages from navigation when disabled in settings
Currently the pages menu entry stays in the navigation,
even if the feature is disabled in the settings and
renders a page telling you that the feature is disabled.
It doesn't allow you to turn the feature on, so it can
probably just be removed.

This also allows to remove navigation items from a
project that don't make sense in the context of
that project.

Changelog: changed