Add willAutoCloseByMergeRequest field to work items API
Added willAutoCloseByMergeRequest to work items
development widget. Field indicates if the work item will
be closed when the listed merge requests in closingMergeRequests
are merged. This can be disabled at the project level. Also
returns false if work item is already closed or if no open
merge request is associated with the work item

Changelog: added