Create bot users with indefinite group and project membership
When a Resource (Group or Project) Access Token is created for, the
following occurrs:

- a User is created (a bot)
- a PAT is created for the bot (known as a Group or Project Access Token)
- a Member record is created to associate the bot and the resource

Previously, the membership would expire when the token expired. When
the Membership expired, the `RemoveExpiredMembersWorker` would destroy
the Member and bot User records. Expired tokens would no longer have
any association with the resource, making auditing and security reviews
more challenging.

Now, when the Feature Flag
`retain_resource_access_token_user_after_revoke` is enabled for a
Resource, new Resource Access Tokens will have Member records that do
not expire. When the token expires, the Member and User records are
retained. (In a different MR these inactive tokens are displayed in
Group and Project Settings).

This is part of https://gitlab.com/gitlab-org/gitlab/-/issues/462217