Fix Content-Disposition not working for Azure in API download
In GitLab 17.1
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/155013
introduced a fix that attempted to send the
`response-content-disposition` query parameter to object storage
providers. gitlab-fog-azure-rm expects this query parameter to be a
string instead of a symbol, however.

To fix this issue and be consistent with the CarrierWave downloads via
Rails controllers, set `response-content-disposition` and
`response-content-type` to strings.

Changelog: fixed