Early return from Subscriptions#new without user
When the user has not signed in already, we want to redirect them
straight to the registration path when the try to buy a subscription.
Previously, this was raising errors as we were redirecting, but we were
also executing all the code in the #new method (including calls on
current_user, which was nil).

Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/472207
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/159356