Track users' access from PIPL-covered countries
For each user accessing GitLab.com, the access is tracked if it is
coming from PIPL-covered countries (by checking HTTP_CF_IPCOUNTRY
request env). If the access is coming from a non PIPL-covered country
and a previous access from a PIPL-covered country was tracked for the
same user, the access count from PIPL-covered countries is reset for
the user. This tracking is only done once every 24 hours for each user.

Further, if the user is determined to be subject to PIPL, a background job is
queued to check if the user belongs to any paid namespace. This is done because
users belonging to paid namespaces are exempt from actions taken to ensure PIPL
compliance.