Load placeholder references from Redis into PG
https://gitlab.com/gitlab-org/gitlab/-/issues/443554 introduced a single
table to contain all details of placeholder user contributions.

We will record 1 row of data per imported record that is associated with
a user, for every importer.

Our table design has the benefit of being simple, but attempting to
write each row to the table seems a bit naïve and is likely to lead to
scalability problems at times.

Failures to write this data to PostgreSQL will lead to a kind of data
loss for customers.

This change writes the placeholder user contribution data to Redis first
which is later collected and batch loaded to the PostgreSQL table.

https://gitlab.com/gitlab-org/gitlab/-/issues/467511