Exclude scripts from coherence check when publishing npm packages
When publishing a npm package, we perform a coherence validation in which we compare the values of "name", "version" & "scripts" properties of the package between the request payload and the package.json file. If they are not the same, the package isn't published. In this MR, we exclude the "scripts" property from this validation as we see many false mismatch reported by users, and the tools used to publish a package could perform many auto-corrections to the property.

Changelog: fixed