Rename flag to importer_user_mapping
This refactors the flag name so that it will be a flag for the core
user mapping feature, to allow 1 flag per importer we roll the feature
out to.

- `importer_user_mapping`, applies to the core user mapping feature
- Have one flag per importer we roll user mapping out to, example
  `bulk_import_user_mapping`. We would always additionally check the
  flag state of `importer_user_mapping` with these flag checks.

This lets us globally switch user mapping off for all importers, if we
need to (could be handy if we find user mapping strains production
services while we’re evaluating it) but still control the rollout to
individual importers in a granular way.