SourceUserMapper return real user only if accepted
In the new importer user mapping, we initially assign the imported user
contributions to placeholder users rather than the real users.

Later, we ask the real user to approve the imported contributions as
theirs, and we remap the imported contributions from the placeholder
user to the real user.

`Import::SourceUserMapper` is the class responsible for returning the
user record to associate with an imported user contribution. It has
existing logic to return the real user if one was previously assigned to
the `Import::SourceUser` record. But there are states within that record
where the real user has not accepted the imported contributions as
theirs.

This change ensures that the real user is only returned if that user has
approved a mapping previously.

https://gitlab.com/gitlab-org/gitlab/-/issues/443533