Do not link to housekeeper feedback issue anymore instead use label
This updates the default commit message and MR description for
housekeeper MRs to no longer include a link to the feedback issue.

This linked feedback issue has become too difficult to load as it has
hundreds of mentions. It's not really serving it's purpose and a label
and a ping is likely to be just as good for communicating feedback.