Fix issue in RTE related to adding text before a mention
If you start the input by tagging (@somebody) someone, and then
relocate the cursor to before the @, you could type to insert text,
but couldn't delete or otherwise interact with it (backspace and left
arrow keys don't respond). When switching to plain-text, anything
typed before the @ disappeared. This commit fixes the issue.