Fix invalid X-Next-Page header for the first page
Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/471566

**Problem**

Code responsible for the pagination of the first page always assumes
that the second page exists. As a result, it sets a X-Next-Page even if
there is no second page.

**Solution**

Fix the calculation of `total_pages` and `next_page` values.

Changelog: fixed