Lint against outdated usages of `helpPagePath`
As docs pages get moved around or deleted, docs links that we display in
the product might become outdated without us noticing.
This adds an ESLint rule that ensures `helpPagePath` always points to
files and anchors that still exist.
In future iterations, we should add a similar rule to check usages of
the `HelpPageLink` Vue component, and a Rubocop rule for the backend
`help_page_path` helper.