Make Elasticsearch client work with latest gem updates
https://github.com/elastic/elasticsearch-ruby/pull/1360 added a
Elasticsearch version validation step, but unfortunately this causes
the client to fail with OpenSearch. Patch the client to maintain
compatibility.

Update the tests to fix test failure issues present in the latest
gem. The tests were failing because elasticsearch-transport attempted
to inspect the raw string for GZIP headers, but the stubbed response
was a Ruby array so this failed.