Jira Cloud App not editable unless active
This change makes the Jira Cloud App integration only editable when the
integration has been activated, as we disable the ability for this
integration to be enabled and disabled in the normal way.

This means that when a user sees a disabled Jira Cloud App integration
they understand that they need to link a relevant group within the
app configuration first before the can enable the integration.

It also clarifies this in our docs, and within the integration form.

This change is behind the `enable_jira_connect_configuration` feature
flag.

https://gitlab.com/gitlab-org/gitlab/-/issues/460663