Fix publishing npm package with custom root folder name
When publishing a npm package that its archive has been bundled with a custom root folder name, an error is raised stating that package.json not found. In this MR, we fix that by validating the archive as long as it has package.json file in its root regardless of the root folder name.

Changelog: fixed