Fix work item parent links namespace_id
This fixes records where we set the wrong namespace_id. The wrong
namespace_id came from records that were we updated the references of
WorkItemParentLinks and set the namespace_id to project.namespace_id
instead of project.project_namespace_id.

The bug got introduced in 17.1 with
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/155113 and fixed
in 17.2 with
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/157332.

The problem with using the wrong namespace_id is that the record would
not get properly deleted when the project gets deleted. This leads to
orphaned records.

Changelog: fixed