Make `manual:e2e-test-pipeline-generate` always available
The automated `e2e:package-and-test-ee` might only run in tier 3
pipelines, and in some cases we might want to run it earlier.
While this can be done by using `pipeline:run-all-e2e` label,
we'll need to trigger a new pipeline after the label has been applied.

Simply always having a way to manually trigger it is just simpler and
easier.

Note that the manual actions are always allowed to fail though.
Even if the manual `follow-up:e2e:package-and-test-ee` failed,
it'll not be shown as an error in the pipeline, but a warning.

After the pipeline reached tier 3, we run the automated one,
which is not allowed to fail, therefore blocking from merging
if it's failing.