Auto close issues for group level issues defaults to false
Initially we assumed that auto closing issues via merge request
was enabled for issues that exist at the group level. Now we
want to default to disabled if an issue exists at the group
level. In the future, we could add a setting to groups to add
the auto close feature. Better this way, as the other way, there's
no way to disable it.

Changelog: changed