Add allow_git_http_fetch_writes FF
Since the `disable_git_http_fetch_writes` FF does not define
a feature flag actor, it is currently not possible to increment
the fetch project statistics without enabling or disabling it
for all projects.

This commit introduces the `allow_git_http_fetch_writes` FF
that allow us to control this and unskip the statistics
for specific projects.