Remove `.ordered` scope from sbom-to-vulnerability ingestion mapping
During review for MR 160015, it was [mentioned][0] that removing the
ordering by severity could simplify the database query.

It doesn't appear that order is explicitly required for this code. The
tests were relying on the order, but it appears that was just taking
advantage of the order for testing convenience, not because it is a
product requirement

There was some further discussion in [this thread][1] that also indicates the order can be removed.

I split this out to a separate MR as [MR 160015][2] is not really near
a merge-able state

[0]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/160015#note_2010753091
[1]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/160015#note_2013921133
[2]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/160015

---

MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/160693
Related to https://gitlab.com/gitlab-org/gitlab/-/issues/472760