Duo Chat Retry Attempts Update
The retry count for Llm::CompletionWorker has been decreased to 3.

This means that if the worker fails to complete a job, it will be
retried up to 3 times before being considered a failure.

This change is likely to improve the reliability of the worker and
ensure that failing jobs do not stay in the retry queue as long.

Changelog: changed
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/160476
EE: true