Fix checkpointing API taking primary key id from client
The previous implementation had a bug where `declared_params` was
including the `id` param which was only meant to be in the URL and only
meant for looking up the workflow. But since this was being used to
create the `Checkpoint` the API was only able to create 1 checkpoint for
workflow because the 2nd one would have a duplicate primary key error.