Inline helper methods to the npm endpoint classes
This is a follow-up to
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/154577

Like the earlier MR, this is also a refactoring to remove the `if
endpoint_scope...` checks in the code. For this MR, as the title says,
we're removing methods in the shared helper module into inline helper
methods in the individual endpoint (project, group, instance) classes.