Replace stylesheet_link_tag_defer with universal_stylesheet_link_tag
The Rails helper stylesheet_link_tag_defer adds the attribute
media set to all to stylesheet link tags. In the past, we used this
as a feature to 'defer' loading these link tags, compared to the ones
which had a media set to print. This behavior was removed and this
commit cleans up the helper stylesheet_link_tag_defer by replacing it
with the new helper universal_stylesheet_link_tag.