Move resource group to control release environments pipelines
Use resource group at the pipeline level instead of job level, because
of the following reasons:

- Make sure there is only one RE pipeline runs at a time. At job level,
there is still a chance that a job of a later pipeline runs before the
previous pipeline.
- We cannot use resource group at the QA jobs to allow parallelism, so
the pipelines just run without waiting for the QA jobs to finish. This
both mixes up the versions and overloads the instance since many
deployments can happen right after each other.