Mark migration spec as allowed to be slow
Migration spec has been reported as slow but the spec is pretty
minimal and straightforward.  After consulting with a test domain
expert it was agreed that this spec could be allowed to be slow,
especially given it will be removed in the near future as it is nearly
a year old.