Introduce AutoFlow client to send arbitrary events
This change set introduces AutoFlow for the first time. The KAS client
now implements a method (`send_autoflow_event`) to send arbitrary cloud
events to AutoFlow.

The implementation is behind the `autoflow_enabled` feature flag. For
this to work it's also required to properly setup KAS with AutoFlow,
which is also disabled by default.