Hide pending todos authored by banned users
Content by banned users is "hidden" from non-instance-admin users. This
change brings the behavior of todos in-line with this behavior and makes
sure abusive content does not remain in the todo lists of any mentioned
users. This does not apply to done items (i.e. that a user has seen) to
avoid potential confusion.

Changelog: changed