Issue closing pattern works with work items
You can set an issue to be automatically closed
when a merge request is merged. For this you need to
use the closing pattern and an issue reference.
This MR allows you to use the closing pattern together
with a work item reference (by URL) at both project and
group level

Changelog: changed