Do not attempt to upsert existing cvs scanners
The continuous vulnerability scanning feature attempts to insert into
the vulnerability scanner table frequently, and holds onto locks as a
result. Due to the frequency, this can slow down transactions that are
waiting on locks to release. To address this, a select statement is made
to find all pre-existing scanners first. Any scanners found are removed
from the bulk insert operation to minimize the chance of a row becoming
locked.

Changelog: performance