Remove CSS-in-Js before running utils comparison script
The `compare_css_util_classes` script has a tendency to fail when the
CSS-in-Js was built prior to running it. This has caused a broken
`master` after we introduced a `prelint:stylelint` NPM script that
builds Tailwind before running Stylelint, which also happens to run
before the comparison script in the `static-analysis` CI job.

We are therefore adding a `prelint:tailwind-utils` NPM script to ensure
that the CSS-in-Js always gets removed before executing the script.