Move license creation via license key/file into a service class
Prior to this change, the logic used to create a new `License` when
applying a license key or uploading a license file was handled directly
in the `LicensesController`. The logic is already too complex for a
controller. With upcoming changes, the logic needs to be modified which
increased the complexity more.

This change moves the logic to create a `License` when a license key is
applied or a license file is uploaded to a dedicated service. This
simplifies the controller specs and makes extending the logic easier and
better maintainable.