Allow to edit files from a commit or tag with WebIDE
Single file editor and other UI editing actions need
to make sure the ref is on top of the branch. WebIDE
doesn't have that limitation and allows editing in such
situation, by applying a new commit onto a newly created
branch. It was possible with a shortcut. This change
makes it possible with the Edit button, while still
showing a fork suggestion when relevant.

Changelog: fixed