Fix missing dependency project dropdown on narrow screen width
This makes sure the dependency project dropdown is still visible
on narrow screen widths by showing a clickable project icon, similar to
the dependency location count. In addition we've added the count next
to the icon. On wider screen widths, the project icon is now also
displayed next to the label "x projects". Finally, if `locationCount`
or `projectCount` is invalid, we do not show the respective
dropdown at all.

Changelog: fixed
EE: true