Shift anchor icon to the left for headers inside a summary tag
When a user adds an h1, h2, h3, h4, h5 or h6 to their markdown/RTE
content, we automatically render an icon link to the left of it that
navigates to the skiplink for that header.

When the user has placed this header inside a `<summary>` tag, the
skiplink and the summary toggle currently clash. This MR shifts the link
icon to the left in this case so there is room for both.

Changelog: fixed