Make Workhorse and QA CI patterns depend on version.yml
When https://gitlab.com/gitlab-org/gitlab/-/merge_requests/157451 was
backported to `16-11-stable-ee`, we noticed that none of the Workhorse
CI tests ran. We would have noticed that the bump to Ruby 3.1.5 would
have failed in CI since no image with Go 1.20 and Git 2.36 was built
with Ruby 3.1.5. The following merge requests were needed to build
with this combination:

* https://gitlab.com/gitlab-org/gitlab-build-images/-/merge_requests/836
* https://gitlab.com/gitlab-org/gitlab-build-images/-/merge_requests/837