Ensure PostReceive cache clear always happens
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/161673
simplified the cache clear mechanism in PostReceive, but if the
exclusive lease cannot be obtained we should still clear the cache
just in case to give the job a chance to see the latest
branches/tags. There may be some problem with the lease (e.g. abruptly
killed job). This commit also fixes the spec to raise the exception by
`in_lock`.