Properly handle empty repository.ff_merge in FromTrainRef merge strategy
The Gitaly API returns an empty response instead of an error if the ref
update portion of the fast-forward merge fails due to a race with
another ref upate. This could cause a merge request to incorrectly
appear merged if the target branch got updated while merge train was
performing a fast-forward merge.

EE: true
Changelog: fixed