Project Admin - Fix delete project after email message
This change adds logic to update HTML validation
for Inactive project deletion form to check
field validity whenever the valid state changes
in Vue rather than when the field itself is updated.

This fixes the bug that in certain scenarios the
Vue validity would update but the HTML would not.
This would block the HTML form to be submitted.

Previously, the HTML validity was only updated
when the field itself was updated in Vue via a
@change event.  However, the fields validity are
related to each other and thus sceanrios were missed
and the HTML validity never got updated.

Changelog: fixed