Optimize ImportUserCreator returning import user
When working from the same namespace object, the service would not find
the import user and would attempt to create it again.

For example:

```ruby
namespace = Group.first

Gitlab::Import::ImportUserCreator.new(portable: namespace).execute
```

Would result in an import user being created. Calling the line again:

```ruby
Gitlab::Import::ImportUserCreator.new(portable: namespace).execute
```

would result in an attempt to create a new import user, which would be
rolled back.

Calling `#reset` on the namespace means the service correctly finds the
import user if it exists, but it introduces a new query.

This change also caches the import user for the duration of the request.
In Sidekiq, this will be the duration of the Sidekiq job.