Implement Duo Code Review evaluation endpoint
This endpoint allows LLM evaluation tools to get the model's
response using the prompt and model we use for Duo Code Review.

We need this because we don't really need to evaluate the entire
feature including the review publishing and other stuff. The LLM
evaluation tools will also be sending the diff, new_path and hunk
to be reviewed as inputs.

This changes the prompt class as well to accept diff, new_path
and hunk. The prompt needs to change since we won't be generating
the new and old hunk lines but will be reviewing the entire
raw hunk instead.

This is still behind the `ai_review_merge_request` feature flag.