Add warning when project with LFS files is missing .gitattributes
Related to https://gitlab.com/gitlab-org/gitlab/-/issues/429467 we have
found some projects got into a state where there were git-lfs files in
the project but the project was missing the `.gitattributes` file. This
is effectively a corrupted project as the `git-lfs` plugin won't know
about any of these files when cloning or pushing and it could lead to a
lot of confusion. We are adding a warning banner on the project page for
such projects to help save time troubleshooting if a project ends up in
this state.