Remove `Security::ReportFetchService` and refactor tests
With the completion of https://gitlab.com/gitlab-org/gitlab/-/merge_requests/161512
we no longer have any frontend code which relies on the `report`
structure, so we can safely remove it from the dependencies response. We
will continue to use it for the project dependencies export to avoid
introducing a breaking change. This change also refactors
`dependencies_controller_spec.rb` to prepare it for the removal of the
`project_level_sbom_occurrences` feature flag.

Changelog: changed
EE: true