Hide invalid closing issues on merge request
The merge dialog of a merge request will display a message
indicating that a list of issues will be closed when the
merge request is merged. This list was not accurate as it
would not take into account the issue's project setting
where auto closing of issues can be disabled. This commit
hides issues that won't actually be closed.

Changelog: fixed