Allow Redis password to be specified in Workhorse URL
Previously the Workhorse config allowed a Redis URL to be used, but
the password component had to be specified inside the `Password`
config field. However, it's common for the URL to contain both the
username and password, so we should allow that. The `Password` field
takes precedence if both are specified, however.

Changelog: changed