Explicitly state returning columns for p_ci_builds
This new PR assigns auto populated columns:
https://github.com/rails/rails/pull/48241

However, a column is identified as auto-populated if it contains
default value with nextval function.

The id column of p_ci_builds table is auto-populated via a trigger
and it's not identified as an auto-populated column by Rails.

Let's explicitly return ['id'] in a function that expects a list
of auto-populated columns